python-rtmbot-master/requirements.txt
algorithmia==1.0.5
backports.functools-lru-cache==1.5
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
enum34==1.1.6
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.4
pandas==0.23.4
pyjson==1.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.11
recv==0.2.0
requests==2.20.1
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
slackclient==1.0.1
subprocess32==3.5.3
tabulate==0.7.5
urllib3==1.24.1
websocket-client==0.54.0
